--- src/main/java/nextstep/courses/domain/Course.java
@@ -1,66 +1,0 @@
-package nextstep.courses.domain;
-
-import java.time.LocalDateTime;
-
-public class Course {
-    private Long id;
-
-    private String title;
-
-    private Long creatorId;
-
-    private final Sessions sessions;
-
-    private final LocalDateTime createdAt;
-
-    private final LocalDateTime updatedAt;
-
-    public Course(String title, Long creatorId, LocalDateTime now) {
-        this(0L, title, creatorId, new Sessions(), now, null);
-    }
-
-    public Course(
-            Long id,
-            String title,
-            Long creatorId,
-            Sessions sessions,
-            LocalDateTime createdAt,
-            LocalDateTime updatedAt) {
-        this.id = id;
-        this.title = title;
-        this.creatorId = creatorId;
-        this.sessions = sessions;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
-    }
-
-    public void addSession(Session session) {
-        sessions.add(session);
-    }
-
-    public int sessionCount() {
-        return sessions.count();
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public Long getCreatorId() {
-        return creatorId;
-    }
-
-    public LocalDateTime getCreatedAt() {
-        return createdAt;
-    }
-
-    @Override
-    public String toString() {
-        return "Course{" + "id="
-                + id + ", title='"
-                + title + '\'' + ", creatorId="
-                + creatorId + ", createdAt="
-                + createdAt + ", updatedAt="
-                + updatedAt + '}';
-    }
-}
