--- src/main/java/nextstep/courses/domain/CoverImage.java
@@ -1,11 +1,0 @@
-package nextstep.courses.domain;
-
-public class CoverImage {
-    private final ImageFile imageFile;
-    private final ImageDimension dimension;
-
-    public CoverImage(String filename, long fileSize, int width, int height) {
-        this.imageFile = new ImageFile(filename, fileSize);
-        this.dimension = new ImageDimension(width, height);
-    }
-}
