--- src/main/java/nextstep/courses/domain/ImageType.java
@@ -1,16 +1,0 @@
-package nextstep.courses.domain;
-
-public enum ImageType {
-    GIF,
-    JPG,
-    PNG,
-    SVG;
-
-    public static ImageType from(String extension) {
-        String upper = extension.toUpperCase();
-        if (upper.equals("JPEG")) {
-            return JPG;
-        }
-        return valueOf(upper);
-    }
-}
