--- src/main/java/nextstep/courses/domain/SessionStatus.java
@@ -1,17 +1,0 @@
-package nextstep.courses.domain;
-
-public enum SessionStatus {
-    PREPARING("준비중"),
-    RECRUITING("모집중"),
-    CLOSED("종료");
-
-    SessionStatus(String description) {
-        this.description = description;
-    }
-
-    private final String description;
-
-    public boolean canEnroll() {
-        return this == RECRUITING;
-    }
-}
