+++ src/main/java/nextstep/courses/domain/course/CourseRepository.java
@@ -1,0 +1,7 @@
+package nextstep.courses.domain.course;
+
+public interface CourseRepository {
+    int save(Course course);
+
+    Course findById(Long id);
+}
