+++ src/main/java/nextstep/courses/domain/course/Sessions.java
@@ -1,0 +1,25 @@
+package nextstep.courses.domain.course;
+
+import java.util.ArrayList;
+import java.util.List;
+import nextstep.courses.domain.session.Session;
+
+public class Sessions {
+    private final List<Session> values;
+
+    public Sessions() {
+        this(new ArrayList<>());
+    }
+
+    public Sessions(List<Session> values) {
+        this.values = new ArrayList<>(values);
+    }
+
+    public void add(Session session) {
+        this.values.add(session);
+    }
+
+    public int count() {
+        return this.values.size();
+    }
+}
