+++ src/main/java/nextstep/courses/domain/image/ImageType.java
@@ -1,0 +1,16 @@
+package nextstep.courses.domain.image;
+
+public enum ImageType {
+    GIF,
+    JPG,
+    PNG,
+    SVG;
+
+    public static ImageType from(String extension) {
+        String upper = extension.toUpperCase();
+        if (upper.equals("JPEG")) {
+            return JPG;
+        }
+        return valueOf(upper);
+    }
+}
