+++ src/main/java/nextstep/courses/domain/session/SessionPeriod.java
@@ -1,0 +1,21 @@
+package nextstep.courses.domain.session;
+
+import java.time.LocalDate;
+
+public class SessionPeriod {
+    private final LocalDate startDate;
+    private final LocalDate endDate;
+
+    public SessionPeriod(LocalDate startDate, LocalDate endDate) {
+        validatePeriod(startDate, endDate);
+        this.startDate = startDate;
+        this.endDate = endDate;
+    }
+
+    private void validatePeriod(LocalDate startDate, LocalDate endDate) {
+        if (endDate.isBefore(startDate)) {
+            throw new IllegalArgumentException(
+                    String.format("종료일은 시작일 이후여야 합니다. (시작: %s, 종료: %s)", startDate, endDate));
+        }
+    }
+}
