--- src/test/java/nextstep/courses/domain/EnrollmentsTest.java
@@ -1,38 +1,0 @@
-package nextstep.courses.domain;
-
-import static org.assertj.core.api.Assertions.*;
-
-import java.time.LocalDateTime;
-import org.junit.jupiter.api.DisplayNameGeneration;
-import org.junit.jupiter.api.DisplayNameGenerator;
-import org.junit.jupiter.api.Test;
-
-@DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
-class EnrollmentsTest {
-
-    @Test
-    void add_정상입력_성공() {
-        var e1 = new Enrollment(1L, 1L, LocalDateTime.now());
-        var e2 = new Enrollment(1L, 2L, LocalDateTime.now());
-        var enrollments = new Enrollments();
-
-        assertThatCode(() -> {
-                    enrollments.add(e1);
-                    enrollments.add(e2);
-                })
-                .doesNotThrowAnyException();
-    }
-
-    @Test
-    void add_중복신청_예외발생() {
-        var enrollment = new Enrollment(1L, 1L, LocalDateTime.now());
-        var enrollments = new Enrollments();
-
-        assertThatThrownBy(() -> {
-                    enrollments.add(enrollment);
-                    enrollments.add(enrollment);
-                })
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이미 수강 신청한 강의입니다.");
-    }
-}
